+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/icon/package-info.java
@@ -20,6 +20,7 @@
  * Package configuration values:
  * <ul>
  * <li>{@link is.codion.swing.common.ui.icon.Icons#ICON_SIZE}
+ * <li>{@link is.codion.swing.common.ui.icon.Icons#ICON_COLOR}
  * </ul>
  */
 @org.jspecify.annotations.NullMarked